requirements.txt
mysql-connector-python==8.0.26
requests==2.26.0
